services/face-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
insightface==0.7.3
numpy==1.24.4
Pillow==10.1.0
boto3==1.29.0
python-multipart==0.0.6
onnxruntime>=1.17.1
opencv-python-headless==4.8.1.78
scipy==1.11.4
